+++ CHANGELOG.md
@@ -43,12 +43,16 @@
 
 - [#3720](https://github.com/ChainSafe/forest/pull/3720) Implement the
   `Filecoin.GetParentMessages` lotus-compatible RPC API.
+- [#3726](https://github.com/ChainSafe/forest/pull/3726) Implement the
+  `Filecoin.StateMinerFaults` lotus-compatible RPC API.
 - [#3735](https://github.com/ChainSafe/forest/pull/3735) Implement the
   `Filecoin.StateAccountKey` lotus-compatible RPC API.
+- [#3744](https://github.com/ChainSafe/forest/pull/3744) Implement the
+  `Filecoin.StateLookupID` lotus-compatible RPC API.
 - [#3727](https://github.com/ChainSafe/forest/pull/3727) Added glif.io calibnet
   bootstrap node peer
 - [#3737](https://github.com/ChainSafe/forest/pull/3737) Added `--n-tipsets`
-  option to `forest-tool api compre`
+  option to `forest-tool api compare`
 
 ### Changed